pyproject.toml
[project]
name = "study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.15",
    "langchain-deepseek>=1.0.1",
    "langchain-openai>=1.1.12",
    "python-dotenv>=1.2.2",
    "transformers>=5.5.3",
]
